bin/requirements.txt
pandas>=2.2.0
tqdm>=4.66.1
biopython>=1.8.3